+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTablePanel.java
@@ -759,7 +759,6 @@ public class EntityTablePanel extends JPanel {
             .build();
     southPanelSplitPane = Components.splitPane()
             .continuousLayout(true)
-            .resizeWeight(0.35)
             .leftComponent(searchFieldPanel)
             .rightComponent(statusPanel)
             .build();