+++ changelog.md
@@ -4,11 +4,14 @@ Codion Change Log
 ## 0.17.21-SNAPSHOT
 ### is.codion.swing.common.ui
 - Controls.SEPARATOR added, used instead of null to represent separators in Controls instances.
+- HintTextField bug fixed, now adjusts the hint text length to prevent painting outside of bounds.
+- FilteredTable search field now has a minimum size instead of columns.
 ### is.codion.swing.framework.ui
 - EntityPanel bug fixed, table refresh control was always enabled, instead of only when the panel was active.
 - EntityPanel.createControls() added, related refactoring.
 - EntityPanel.createControlsComponent() renamed createControlComponent().
 - EntityPanel.editPanel() and tablePanel() now throw IllegalStateException in case no edit panel or table panel is available.
+- EntityTablePanel south panel split pane resize weight no longer specified, for a more consistent initial search field size.
 
 ## 0.17.20
 ### is.codion.common.model